backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-generativeai
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
python-dateutil==2.8.2
